moved stuff out of flaskApi.py and utils.py into own files
