+++ api-server-flask/api/requirements.txt
@@ -1,0 +1,67 @@
+aiohttp==3.9.3
+aiosignal==1.3.1
+async-timeout==4.0.3
+attrs==23.2.0
+blinker==1.7.0
+cachetools==5.3.2
+certifi==2024.2.2
+cffi==1.16.0
+charset-normalizer==3.3.2
+click==8.1.7
+cloud-sql-python-connector==1.7.0
+colorama==0.4.6
+cryptography==42.0.2
+Deprecated==1.2.14
+diff-match-patch==20230430
+exceptiongroup==1.2.0
+flask==3.0.1
+Flask-Cors==4.0.0
+flask-sqlalchemy==3.1.1
+frozenlist==1.4.1
+google-api-core==2.17.1
+google-api-python-client==2.118.0
+google-auth==2.27.0
+google-auth-httplib2==0.2.0
+google-cloud==0.34.0
+google-cloud-core==2.4.1
+google-cloud-storage==2.14.0
+google-crc32c==1.5.0
+google-resumable-media==2.7.0
+googleapis-common-protos==1.62.0
+greenlet==3.0.3
+gunicorn==21.2.0
+httplib2==0.22.0
+idna==3.6
+importlib-metadata==7.0.1
+iniconfig==2.0.0
+itsdangerous==2.1.2
+Jinja2==3.1.3
+MarkupSafe==2.1.4
+multidict==6.0.5
+mysql-connector-python==8.3.0
+oauth2client==4.1.3
+packaging==23.2
+pluggy==1.4.0
+protobuf==4.25.2
+pyasn1==0.5.1
+pyasn1-modules==0.3.0
+pycparser==2.21
+PyGithub==2.3.0
+PyJWT==2.8.0
+PyMySQL==1.1.0
+PyNaCl==1.5.0
+pyparsing==3.1.1
+pytest==8.1.1
+python-dotenv==1.0.1
+requests==2.31.0
+rsa==4.9
+six==1.16.0
+SQLAlchemy==2.0.25
+tomli==2.0.1
+typing-extensions==4.9.0
+uritemplate==4.1.1
+urllib3==2.2.0
+werkzeug==3.0.1
+wrapt==1.16.0
+yarl==1.9.4
+zipp==3.17.0
